Commit for Assn 9. Tabbed/Modal pages.
Create an app that includes three Tabbed Pages and a Modal Page.
A tabbed page that introduces you with an image
A tabbed page that asks your age and states how many years you could legally drive in Wisconsin
A tabbed page with credits for the creation of the app.
A modal page that has terms of use, accessed from the credits page